commit avant pull

+++ src/components/social/gestion_employes/Supprimer_employe.js
@@ -19,16 +19,22 @@ class SupprimerEmploye extends Component {
     console.log(idEmploye);
     const idSociete = UserService.getSocietyId;
     console.log(idSociete);
-    AxiosCenter.deleteWrapperEmploye(idEmploye)
-      .then((response) => {
-        console.log(response);
-        this.setState({
-          modal: !this.state.modal,
+    const roleUser = UserService.getRole;
+    console.log(roleUser);
+    if (roleUser === "ROLE_SOCIETY") {
+      AxiosCenter.deleteWrapperEmploye(idEmploye)
+        .then((response) => {
+          console.log(response);
+          this.setState({
+            modal: !this.state.modal,
+          });
+        })
+        .catch((error) => {
+          console.log(error);
         });
-      })
-      .catch((error) => {
-        console.log(error);
-      });
+    } else {
+      alert("NON AUTORISE !");
+    }
   };
 
   render() {
